Freeze top-level packages at arbitrary existing versions
We have not updated this image in ages, and when we just
bumped up a couple of
packages (https://github.com/2i2c-org/2i2c-hubs-image/pull/3),
it actually bumped up a *lot* of other packages to newer versions.
This will break a lot of users in the middle of a semester, which
is problematic.

Ideally, we should have 'proper' pinning (like in
https://github.com/CryoInTheCloud/hub-image), but given the lack
of that, I have just specifically pinned the versions to what we
currently have while we work on a longer term solution.